requirements.txt
numpy
opencv-python
onnxruntime
torch
matplotlib
Pillow
pdfplumber
huggingface_hub
shapely
pyclipper
ruamel.yaml
cachetools
six
fastapi
uvicorn
pydantic
mem0ai
python-dotenv
google-generativeai